// src/app/dataset-wizard/enums/dataset.enum.ts
export enum DatasetStatusEnum {
  IN_DRAFT = 'IN_DRAFT',
  IN_PROGRESS = 'IN_PROGRESS',
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
}

export enum DatasetStepEnum {
  DATA_SETTINGS = 'DATA_SETTINGS',
  MAPPING_SETTINGS = 'MAPPING_SETTINGS',
  DATA_VALIDATION = 'DATA_VALIDATION',
  FINALIZE_SAVE = 'FINALIZE_SAVE',
}